separate script from library function
